Convert to TDD style tests.
The current test take a TDD style approach. Let's change the mocha/chai
language to reflect TDD style instead of using BDD. BDD tests will
likely be added later, but the primary development paradigm is TDD, so
let's use tests meant to match that style.

Note that this breaks Coveralls and Travis. The Mocha calls for those
tools is failing due to the "-u" argument. This will be repaired in
another commit.