Begin separating extractors from sanitizers.
It's becoming clear that with some properties there are two unique
processes, extracting and sanitizing. The extractors will usually take a
payload and find the right property based on whatever logic is
necessary. The sanitizers will then clean that value.

This commit separates those two pieces for the `document-uri` property.
I will likely break this into a new structure to make extractors first
class citizens in the app.