Major refactor.
Refactored functionality for effective directive into 4 parts: extract,
validate, sanitize, and get. Effective directive and violated directive
now use this 4 processes. Others will be migrated to this format as
well.

This change was necessary to handle all of the needs of getting the
effective directive. Essentially, you must first extract the data from
the report. Then, depending on the type of data, you must either
sanitize or validate it. Finally, the "get" functionality is what wraps
these subprocesses and produces a final result.

This process works well for effective directives and should be utilized
for others. This is particularly helpful for more complex data where
multiple processes are needed.